requirements.txt
# Python 3.11 required (higher versions may cause issues with hume)

# Core video processing dependencies
pydub>=0.25.1
pillow>=8.0.0
numpy>=1.20.0
httpx>=0.24.0
hume>=0.2.0

# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1